requirements.txt
# AI for Health - Integrated Health Intelligence System
# 综合健康AI系统依赖包

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# LLM and AI frameworks
openai>=1.0.0
anthropic>=0.3.0
langchain>=0.1.0
langgraph>=0.0.1

# Deep learning (for Sports Injury Risk)
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Graph and memory
networkx>=2.8.0

# Data processing
pydantic>=2.0.0

# Async support
asyncio-compat>=0.1.0

# Logging and utilities
python-dotenv>=1.0.0

# Optional: For visualization
matplotlib>=3.5.0
plotly>=5.0.0

# Optional: For API deployment
fastapi>=0.100.0
uvicorn>=0.23.0

